Refactor MerakiAPIClient to separate data orchestration
- Extracted data fetching orchestration logic from `MerakiAPIClient` to new `DataFetchManager` class.
- Updated `MerakiDataUpdateCoordinator` to use `DataFetchManager`.
- Cleaned up `MerakiAPIClient` to focus on API endpoint access.
- Updated `ClientFetcher` and `DeviceFetcher` to use public `run_with_semaphore`.
- Refactored and updated tests to reflect architecture changes.
- Added `has_dashboard` property to `MerakiAPIClient` to safe-guard access.
