+++ custom_components/meraki_ha/const.py
@@ -12,11 +12,10 @@ from typing import Final
 # Re-export constants from sub-modules
 from .const_api import *  # noqa: F403
 from .const_conf import *  # noqa: F403
-from .const_device import *  # noqa: F403
-from .const_platform import *  # noqa: F403
-from .const_device import *  # noqa: F403
-from .const_sensor import *  # noqa: F403
 from .const_data import *  # noqa: F403
+from .const_device import *  # noqa: F403  # noqa: F403
+from .const_platform import *  # noqa: F403
+from .const_sensor import *  # noqa: F403
 
 DOMAIN: Final = "meraki_ha"
 """Domain for the component."""
