+++ tests/core/api/endpoints/test_appliance.py
@@ -39,7 +39,7 @@ def api_client(hass, mock_dashboard, coordinator):
     """Fixture for a MerakiAPIClient instance."""
     with patch("meraki.DashboardAPI", return_value=mock_dashboard):
         client = MerakiAPIClient(
-            hass=hass, api_key="test-key", org_id="test-org", coordinator=coordinator
+            hass=hass, api_key="test-key", org_id="test-org"
         )
         client.dashboard = mock_dashboard
         yield client
